[fix] 소켓 연동 구현 중..

+++ app/src/main/java/com/tosslab/jandi/app/events/entities/TopicDeleteEvent.java
@@ -4,19 +4,19 @@ package com.tosslab.jandi.app.events.entities;
  * Created by Steve SeongUg Jung on 15. 4. 9..
  */
 public class TopicDeleteEvent {
-    private final int teamId;
-    private final int id;
+    private final long teamId;
+    private final long id;
 
-    public TopicDeleteEvent(int teamId, int id) {
+    public TopicDeleteEvent(long teamId, long id) {
         this.teamId = teamId;
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public int getTeamId() {
+    public long getTeamId() {
         return teamId;
     }
 }
